--- src/schemas/places/delete.js
@@ -1,7 +1,0 @@
-const { z } = require("zod");
-
-const deletePlaceSchema = z.object({
-  params: z.object({ id: z.string().uuid() }),
-});
-
-module.exports = { deletePlaceSchema };
