--- src/schemas/places/get.js
@@ -1,7 +1,0 @@
-const { z } = require("zod");
-
-const getPlaceSchema = z.object({
-  params: z.object({ id: z.string().uuid() }),
-});
-
-module.exports = { getPlaceSchema };
