+++ src/validation/places/delete.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const deletePlaceSchema = z.object({
+  params: z.object({ id: z.string().uuid() }),
+});
