+++ src/validation/places/get.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const getPlaceSchema = z.object({
+  params: z.object({ id: z.string().uuid() }),
+});
