+++ src/validation/users/get.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const getUserSchema = z.object({
+  params: z.object({ id: z.string().uuid() }),
+});
