+++ src/validation/users/post.ts
@@ -1,6 +1,6 @@
-const { z } = require("zod");
+import { z } from "zod";
 
-const createUserSchema = z.object({
+export const createUserSchema = z.object({
   body: z.object({
     name: z.string(),
     email: z.string().email(),
@@ -8,11 +8,9 @@ const createUserSchema = z.object({
   }),
 });
 
-const userLoginSchema = z.object({
+export const userLoginSchema = z.object({
   body: z.object({
     email: z.string().email(),
     password: z.string(),
   }),
-})
-
-module.exports = { createUserSchema, userLoginSchema };
+});
